+++ yggdrasil_decision_forests/metric/report_test.cc
@@ -156,7 +156,7 @@ TEST(Report, HtmlReportAnomalyDetection) {
   // Configure the evaluation.
   utils::RandomEngine rnd;
   proto::EvaluationOptions option;
-  option.set_task(model::proto::Task::ANOMALY_DETECTION);
+  option.set_task(model::proto::Task::CLASSIFICATION);
 
   // Initialize.
   proto::EvaluationResults eval;
@@ -168,10 +168,22 @@ TEST(Report, HtmlReportAnomalyDetection) {
 
   // Add some predictions.
   pred.mutable_anomaly_detection()->set_value(0.5);
-  ASSERT_OK(AddPrediction(option, pred, &rnd, &eval));
+  // It is not possible to add an anomaly detection prediction directly.
+  EXPECT_FALSE(AddPrediction(option, pred, &rnd, &eval).ok());
+
+  model::proto::Prediction classification_pred;
+  ASSERT_OK(ChangePredictionType(model::proto::Task::ANOMALY_DETECTION,
+                                 model::proto::Task::CLASSIFICATION, pred,
+                                 &classification_pred));
+  classification_pred.mutable_classification()->set_ground_truth(1);
+  ASSERT_OK(AddPrediction(option, classification_pred, &rnd, &eval));
 
   pred.mutable_anomaly_detection()->set_value(0.1);
-  ASSERT_OK(AddPrediction(option, pred, &rnd, &eval));
+  ASSERT_OK(ChangePredictionType(model::proto::Task::ANOMALY_DETECTION,
+                                 model::proto::Task::CLASSIFICATION, pred,
+                                 &classification_pred));
+  classification_pred.mutable_classification()->set_ground_truth(1);
+  ASSERT_OK(AddPrediction(option, classification_pred, &rnd, &eval));
 
   // Finalize.
   ASSERT_OK(FinalizeEvaluation(option, label_column, &eval));